# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Impreza Host Public API
  description: |
    Official REST API for Impreza Host clients. Manage domains, hosting, VPS, email services, and place orders programmatically.

    ## Authentication

    All requests require two headers:
    - `X-API-Key` — Your API key (starts with `imp_`)
    - `X-API-Secret` — Your API secret (shown once when generated)

    Generate keys in your [Impreza Account → API Management](https://portal.imprezahost.com/imprezaapi.php).

    ## IP Whitelist

    Your request IP must be whitelisted before making API calls. Add IPs in the API Management page.

    ## Rate Limiting

    Default: **60 requests per minute** per API key. Response headers:
    - `X-RateLimit-Limit` — Max requests per minute
    - `X-RateLimit-Remaining` — Remaining requests in current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets

    ## Destructive Operations

    Irreversible actions (reinstall, terminate, rollback) require `"confirm": true` in the request body.

    ## Errors

    All errors follow a consistent format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable description"
      },
      "meta": { "request_id": "req_abc123", "timestamp": "2026-03-30T14:00:00Z" }
    }
    ```
  version: 1.0.0
  contact:
    name: Impreza Host Support
    url: https://imprezahost.com
  license:
    name: Proprietary
servers:
...
    description: Production
security:
  - ApiKey: []
    ApiSecret: []
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Your API key (e.g. `imp_a1b2c3d4...`)
    ApiSecret:
      type: apiKey
      in: header
      name: X-API-Secret
      description: Your API secret
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          $ref: '#/components/schemas/Meta'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_REQUEST
            message:
              type: string
              example: "Field 'hostname' is required."
            details:
              type: object
        meta:
          $ref: '#/components/schemas/Meta'
    Meta:
      type: object
      properties:
        request_id:
          type: string
          example: req_a1b2c3d4e5f6g7h8
        timestamp:
          type: string
          format: date-time
          example: "2026-03-30T14:05:00Z"
...
    Order:
      type: object
      properties:
        id:
          type: integer
          example: 890
        order_number:
          type: string
          example: "4829174"
        date:
          type: string
          example: "2026-03-30"
        amount:
          type: number
          example: 15.00
        invoice_id:
          type: integer
          example: 2345
        status:
          type: string
          example: Active
        payment_method:
          type: string
          example: mailin
...
  responses:
...
    Forbidden:
      description: IP not whitelisted or resource not owned
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /products/{id}:
    get:
      tags: [Products]
      summary: Get product details
      description: |
        Returns a single product with full pricing per billing cycle, setup fees,
        custom fields, and configurable options (e.g. OS template) required for ordering.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product details with pricing, custom fields, and configurable options
          content:
            application/json:
              example:
                success: true
                data:
                  id: 15
                  name: "VPS Plan 2"
                  description: "4 vCPU, 8GB RAM, 100GB SSD"
                  type: server
                  group: "VPS Hosting"
                  group_id: 3
                  currency: USD
                  pricing:
                    monthly: { price: 15.00, setup_fee: 0.00 }
                    annually: { price: 144.00, setup_fee: 0.00 }
                  custom_fields:
                    - { id: 1, name: "Hostname", type: "text", required: true }
                  config_options:
                    - id: 5
                      name: "Operating System"
                      type: 1
                      options:
                        - { id: 10, name: "Ubuntu 22.04", pricing: { monthly: 0.00 } }
                        - { id: 11, name: "Debian 12", pricing: { monthly: 0.00 } }
                        - { id: 12, name: "Windows Server 2022", pricing: { monthly: 5.00 } }
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════════════════
  # ORDERS
  # ═══════════════════════════════════════════════════════
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - properties:
                      data:
                        type: object
                        properties:
                          orders:
                            type: array
                            items:
                              $ref: '#/components/schemas/Order'
                          total:
                            type: integer

    post:
      tags: [Orders]
      summary: Create a new order
      description: |
        Creates a new order and pays from account credit balance.
        If the balance is insufficient, returns `409 INSUFFICIENT_BALANCE`.

        Use `GET /products/{id}` to discover available `config_options` (e.g. OS template,
        number of email accounts) and `custom_fields` for the product before placing an order.

        ### Examples

        **VPS with OS template:**
        ```json
        {
          "product_id": 15,
          "billing_cycle": "monthly",
          "hostname": "my-vps",
          "config_options": { "5": "10" }
        }
        ```

        **Titan Email with 5 accounts:**
        ```json
        {
          "product_id": 20,
          "billing_cycle": "annually",
          "domain": "example.com",
          "config_options": { "8": "5" }
        }
        ```

        **Google Workspace with 3 seats:**
        ```json
        {
          "product_id": 25,
          "billing_cycle": "annually",
          "domain": "example.com",
          "config_options": { "12": "3" }
        }
        ```
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product_id, billing_cycle]
              properties:
                product_id:
                  type: integer
                  description: Product ID from Impreza Account
                  example: 15
                billing_cycle:
                  type: string
                  enum: [monthly, quarterly, semiannually, annually, biennially, triennially]
                  example: monthly
                domain:
                  type: string
                  description: Domain for the service (if applicable)
                  example: myserver.com
                hostname:
                  type: string
                  description: Hostname for VPS services
                  example: my-vps
                config_options:
                  type: object
                  description: |
                    Configurable options (e.g. OS template). Object keys are option IDs,
                    values are the selected sub-option IDs. Get available options from `GET /products/{id}`.
                  example: { "5": "10" }
                custom_fields:
                  type: object
                  description: |
                    Custom field values. Object keys are field IDs, values are the field values.
                    Get available fields from `GET /products/{id}`.
                  example: { "1": "myserver.example.com" }
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              example:
                success: true
                data:
                  order_id: 890
                  invoice_id: 2345
                  product: "VPS Plan 2"
                  amount: 15.00
                  status: Active
                  message: "Order created and paid from account balance."
        '400':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Insufficient balance
          content:
            application/json:
              example:
                success: false
                error:
                  code: INSUFFICIENT_BALANCE
                  message: "Insufficient balance. Required: $15.00, Available: $3.50"
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Order details with items
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{id}/upgrade:
    post:
      tags: [Orders]
      summary: Upgrade a service
      description: Creates an upgrade order paid from account credit balance.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service_id, new_product_id, billing_cycle]
              properties:
                service_id:
                  type: integer
                  example: 567
                new_product_id:
                  type: integer
                  example: 25
                billing_cycle:
                  type: string
                  example: monthly
      responses:
        '201':
          description: Upgrade created
        '409':
          description: Insufficient balance

  # ═══════════════════════════════════════════════════════
  # INVOICES
  # ═══════════════════════════════════════════════════════
...
  /vps/proxmox/{serviceId}/resources:
    get:
      tags: [VPS Proxmox]
      summary: Get resource usage
      description: CPU, memory, disk I/O, network statistics.
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Resource usage
...